Restructure code and update coordinates
