+++ Clash Royale Helper/Clash Royale Helper/cardCollection.txt
@@ -16,12 +16,16 @@ CannonCartCard:5
 CloneCard:3
 DarkPrinceCard:4
 DartGoblinCard:3
+EarthquakeCard:3
+ElectroDragonCard:5
 ElectroWizardCard:4
 EliteBarbariansCard:6
 ElixirCollectorCard:6
+ElixerGolemCard:3
 ExecutionerCard:5
 FireSpiritsCard:2
 FireballCard:4
+FishermanCard:3
 FlyingMachineCard:4
 FreezeCard:4
 FurnaceCard:4
@@ -29,8 +33,10 @@ GiantCard:5
 GiantSkeletonCard:6
 GiantSnowballCard:2
 GoblinBarrelCard:3
+GoblinCageCard:4
 GoblinGangCard:3
 GoblinHutCard:5
+GoblinGiantCard:6
 GoblinsCard:2
 GolemCard:8
 GraveyardCard:5
@@ -64,11 +70,13 @@ PoisonCard:4
 PrinceCard:5
 PrincessCard:3
 RageCard:2
+RamRiderCard:5
 RascalsCard:5
 RocketCard:6
 RoyalGhostCard:3
 RoyalGiantCard:6
 RoyalHogsCard:5
+RoyalRecruitsCard:7
 SkeletonArmyCard:3
 SkeletonBarrelCard:3
 SkeletonsCard:1
@@ -80,6 +88,7 @@ ThreeMusketeersCard:9
 TombstoneCard:3
 TornadoCard:3
 ValkyrieCard:4
+WallBreakerCard:2
 WitchCard:5
 WizardCard:5
 X-BowCard:6
